Revert "Don't judge me, just for tests"
This reverts commit f973db50103dc8f3acee54d3d97ba4bd10aa62a4.

+++ Jenkinsfile
@@ -74,13 +74,12 @@ pipeline{
             steps{
                 checkout scm
                 sh """
-                 rm -rf spark   
-
-                 wget -q "http://d3kbcqa49mib13.cloudfront.net/${env.SPARK}.tgz"
-                 mkdir -p "${env.SPARK_HOME}"
-                 tar zxvf ${env.SPARK}.tgz -C "${env.SPARK_HOME}" --strip-components 1
-                 rm -rf ${env.SPARK}.tgz
-               
+                if [ ! -d "${env.SPARK_HOME}" ]; then
+                        wget -q "http://d3kbcqa49mib13.cloudfront.net/${env.SPARK}.tgz"
+                        mkdir -p "${env.SPARK_HOME}"
+                        tar zxvf ${env.SPARK}.tgz -C "${env.SPARK_HOME}" --strip-components 1
+                        rm -rf ${env.SPARK}.tgz
+                fi
                 """
             }
         }
